components/services/chat-service/pyproject.toml
[tool.poetry]
name = "chat-service"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
jwt = "^1.3.1"
peewee = "^3.17.8"
pyjwt = "^2.10.0"
psycopg2-binary = "^2.9.10"
eventlet = "^0.38.0"
python-socketio = "^5.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
